Module Tuyau

module Sigs : sig ... end

A Protocol.

A communication protocol is a system of rules that allows entities to transmit information. In the case of Tuyau, this kind of information must not be arbitrary. The protocol should only solve communication problems such as routing.

When we talk about a protocol, it's only about a standard which is able to transmit a payload. Interpretation of the payload is not done by the protocol but by the user of this library.

For example, the Transmission Control Protocol (TCP) is a protocol according to Tuyau because it is able to transmit payload without interpreting it. A counter example is the Simple Mail Transfer Protocol (SMTP) which gives an interpretation of the payload (such as EHLO which is different to QUIT).

This difference is important to unlock the ability to compose protocols. An other protocol according to Tuyau is Transport Layer Security (TLS) - which wants to solve privacy and data integrity. Tuyau is able to compose protocols together like TCP ∘ TLS to make a new protocol. From this composition, the user is able to implement Secure Simple Mail Transfer Protocol (SSMTP) or HyperText Transfer Protocol Secure (HTTPS) - both use TCP and TLS.

A Flow.

To be able to do this composition, the protocol must respect an interface: the FLOW interface. It defines an abstract type t and functions like recv or send. These functions give to us the payload. Rules to solve communication problems are already processed internally.

In other terms, from a given FLOW, the user should not handle routing, privacy or data integrity (or some others problems). The user should only be able to process the payload.

Finally, representation of a TCP protocol is a FLOW. VCHAN protocol or User Datagram Protocol (UDP) can be represented by a FLOW. However, TLS is not a flow but a layer on top of another protocol. Composition with it should look like:

val with_tls : (module FLOW) -> (module FLOW) 

From a given FLOW, we wrap it with TLS and return a new FLOW. Such a composition exists also for WireGuard or Noise layers. Tuyau wants to solve this composition by a strict OCaml interface of the FLOW.

Resolution.

Tuyau wants to solve one last problem, resolution of an endpoint. The goal is to make a FLOW from an endpoint given by the developer.

Definition of an endpoint can not fully exist where it depends on the returned FLOW. For example, if we give to you a TCP flow, endpoint should be an IP and a port where the given FLOW is already connected.

However, we agree that the most general (by convention) description of the endpoint is the domain-name. By knowing this, we let the developer to construct an endpoint from a [ `host ] Domain_name.t.

At the end, Tuyau should be able to construct an endpoint from a [ `host ] Domain_name.t. Then, it tries to find a SERVICE according to the given endpoint and initializes a FLOW.

The most abstract definition provided by Tuyau is:

val flow : resolvers -> [ `host ] Domain_name.t -> flow 

Where resolvers is a set of heterogeneous constructors of endpoints given by the developer. The returned value flow is an abstraction of an already initialized communication protocol. From it, the developer can extract send and recv functions (as described into A Protocol).

Conclusion.

Tuyau is a framework which wants to give a few definitions to restrict developers of protocols to an interface FLOW and, by this way, provide them with a set of tools to compose with others protocols and give only one way to resolve an endpoint (whatever its definition).

Tuyau does not make magic and all described processes previously are explicit - composition, resolution, extraction. This last aspect wants to solve a well-known problem of Conduit where nobody can fully understand this framework.

You can start to read the rest of the documentation.

module type S = sig ... end

Composition.

Tuyau does not do something magic as we said into the introduction. Composition of protocols must be done by protocol developer. Tuyau gives interfaces which can be help this composition - but the glue needed must be implemented.

Considering TLS as a layer which can compose with an other protocol, the implementation looks like:

type input
type output
type +'a s

type 'flow with_tls =
  { flow : 'flow
  ; tls : Tls.Engine.state }

module With_tls
    (Flow : Sigs.F with type input = input
                    and type output = output
                    and type +'a s = 'a s)
= struct
  type flow = Flow.flow with_tls
  type endpoint = Flow.endpoint * Tls.Config.client

  ...
end

let with_tls
  :  type edn flow.
     key:edn key
  -> flow Witness.protocol
  -> (edn * Tls.Config.client) key * flow with_tls Witness.protocol
  = fun ~key protocol ->
    match impl_of_protocol ~key protocol with
    | Ok (module Flow) ->
      let module M = With_tls(Flow) in
      let k = key "with_tls" in
      let p = register_protocol ~key:k ~protocol:(module M) in
      k, p
    | Error err -> failwithf "%a" pp_error err

module Make : functor (Scheduler : Sigs.SCHEDULER) -> functor (Input : Sigs.SINGLETON) -> functor (Output : Sigs.SINGLETON) -> S with type input = Input.t and type output = Output.t and type +'a s = 'a Scheduler.t